Add string len check to HDF output
Added a check to ensure that the HDF output was using the correct string
length for storage. When using chunksize, the string storage was initialized
with the length of the longest string, which might not be the longest total
size. Use the header info from the dbf to set this properly.