Add index column to textsql
I realized after the fact that it would be better to have an index column for
my SQL tables. I added a little logic to add an index primary key to all
tables. There is no keyword switch for this as of now, because most folks are
probably going to want this anyway.